Moving do_history() from easyaccess to do_utils. do_history() still needs to be fixed because it needs to reference a local variable in easyaccess
